% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdqs.R
\docType{package}
\name{WDQS-package}
\alias{WDQS}
\alias{WDQS-package}
\title{WikidataQueryServiceR: An R Wrapper For Wikidata Query Service API}
\description{
This is an R wrapper for the
  \href{https://www.mediawiki.org/wiki/Wikidata_query_service}{Wikidata Query Service}
  (WDQS) which provides a way for tools to query Wikidata via SPARQL.
}
\references{
\itemize{
  \item \href{https://www.mediawiki.org/wiki/Wikidata_query_service/User_Manual}{WDQS User Manual}
  \item \href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/queries/examples}{SPARQL Query Examples} for WDQS
}
}

