% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{WDQSR-deprecated}
\alias{WDQSR-deprecated}
\alias{scrape_example}
\title{Deprecated functions}
\usage{
scrape_example(example_name, ...)
}
\arguments{
\item{example_name}{the names of the examples as they appear on
\href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/queries/examples}{this page}}

\item{...}{ignored (kept for backwards-compatibility)}
}
\description{
Why did I have to go and make things so deprecated?
}
\section{Functions}{
\itemize{
\item \code{scrape_example}: use \link{get_example} instead which employs \link[WikipediR:page_content]{WikipediR::page_content}
}}

