% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdqs.R
\docType{package}
\name{WDQS-package}
\alias{WDQS-package}
\alias{WDQS}
\title{WikidataQueryServiceR: An R Wrapper For Wikidata Query Service API}
\description{
This is an R wrapper for the
\href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service}{Wikidata Query Service}
(WDQS) which provides a way for tools to query Wikidata via
\href{https://en.wikipedia.org/wiki/SPARQL}{SPARQL}.
}
\details{
\href{https://www.mediawiki.org/wiki/Wikidata_query_service}{Wikidata Query Service}
is maintained by \href{https://wikimediafoundation.org/}{Wikimedia Foundation}.
}
\references{
\itemize{
\item \href{https://www.wikidata.org/wiki/Wikidata:A_beginner-friendly_course_for_SPARQL}{A beginner-friendly course for SPARQL}
\item Building a SPARQL query: \href{https://www.wikidata.org/wiki/Help:SPARQL/Building_a_query/Museums_on_Instagram}{Museums on Instagram}
\item \href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/queries/examples}{SPARQL Query Examples} for WDQS
\item \href{http://programminghistorian.org/lessons/graph-databases-and-SPARQL}{Using SPARQL to access Linked Open Data}
by Matthew Lincoln
\item Interesting or illustrative \href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/queries}{SPARQL queries}
for Wikidata
\item Wikidata \href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/2016_SPARQL_Workshop}{2016 SPARQL Workshop}
\item \href{https://www.youtube.com/watch?v=1jHoUkj_mKw}{Wikidata SPARQL Query video tutorial}
by Navino Evans
\item \emph{\href{http://www.learningsparql.com/}{Learning SPARQL}} by Bob DuCharme
\item \href{https://www.mediawiki.org/wiki/Wikidata_query_service/User_Manual}{WDQS User Manual}
}
}
